This isn't much of an update
I was still in the middle of working on my script and decided to record a video of the playspace to use for testing the monsters later.
The video looked messed up for no apparent reason and then my entire project got corrupted. I spent a while trying to fix it before just giving up and making a new branch. I'm just saving this now so I can quickly revert when something goes wrong next.
I don't think we're gonna be testing if this works or not before tomorrow.